Refactor: Split service test suite by domain
